And onceOrIf callback to Connection object
In one of the until attach tests, the connection on(‘connected’) was firing even though it was already connected.  With the recent change to remove that functionality to ensure consistency with other client libraries and the spec, this test fails. So onceOrIf has been added as an undocumented feature to assist with these types of situations.