Avoid problems if conn closed in publish cb for a failed channel
So in channelattach_publish_invalid, the connection is closed in the
publish callback, which fails because the channel name is invalid.

The publish callback is called by failPendingMessages, which is called
by setDetached.

Problem is: the connection closing triggers connectionManager's notifyState
to call setSuspended on the channel, which itself triggers
failPendingMessages (while still inside the original
failPendingMessages) -- and the channel state is not yet 'failed' as
setDetached was calling failPendingMessages before setState.

So: reorder operations so channel state is set before
failPendingMessages, and don't run setSuspended if the channel state is
already detached or failed.