Change MessageQueue#append/prepend to take arrays, not messageQueues
Because ConnectionManager is the only thing to use those methods, and
it's feeding them arrays (more specifically: it's feeding them the
output of a MessageQueue#copyAll, which is an array).

The only reason we haven't noticed this up till now is most browsers
treat `Array.prototype.push.apply([], undefined)` as a noop, but IE8
throws an error. Thank goodness for IE8, eh?