Rewrite resume test to work with max two concurrent http connections
If IE8 thinks it's on a narrowband connection -- using a heuristic that
seems to always be true in a VM -- it limits you to two concurrent
connections per page to the same host.

Problem is, when you have two realtime connections open, with one
sending messages -- as the resume test does -- this results in
deadlocks. Eg maybe both recvs wait for 15 seconds until
one get a heartbeat so the send can jump in, or something. (To make
things worse, the IE network tab shows 'not yet sent because of this
limit' and 'server not yet replied' in the same way -- as "(pending)" --
so by appearances it seems like realtime just took 15 seconds to
respond to a /send request).

Solved by replacing txRealtime with a rest client that doesn't use
up an http slot hanging on a recv. That helped but still wasn't enough
to make it pass consistently, so also changed it to wait for each message to
be received after it's sent.