Fixing issue #205 - Failing test auth_token_expires
The Rest client would fail to connect if the client's clock is ahead of
the server's clock by more than the auth token's TTL. The serverTimeOffset
was calculated only when the queryTime auth param was set, thus any connections
initiated without queryTime=true would not benefit from the serverTimeOffset
mechanism, and the following test from Auth.prototype.authorise would fail:

    if(token.expires === undefined || (token.expires > this.getTimestamp())) {

The proposed solution memoizes the time offset on the Rest object's prototype,
making sure that the offset is calculated once per client, as opposed to
once per request.

Memoization behaviour can be overriden with the queryTime bool parameter, if set.
Thus the offset is calculated in the following cases:

                                 queryTime
                    |  truthy  |   falsy   |   not set
             -------|----------|-----------|-----------
offset===null  true |   yes    |    no     |    yes
              false |   yes    |    no     |    no

When reviewing, please verify that this is the desired behaviour.