Safeguard Realtime Constructor
Ably.Realtime is part of the public api to the whole library. I came across this because the presence docs have an example where the new keyword is not used to instantiate a client. Currently this throws an obscure error. `TypeError: this.connect is not a function(…)`

It seams that `new Ably.Realtime()` should never be called without the new keyword so this safeguard makes sense'