Make getCipher synchronous
Since channels.get() is sync, having getCipher be async causes
problems in node if eg you try to publish immediately after getting a
channel, since the encryption might not have initialized yet. Added test
for this.

This means we're calling node's randomBytes synchronously, but in practice this
really should not be a problem - see commit message and comment in
https://gitlab.pluribusgames.com/mirrors/io-js/commit/f68a116c3c061151ab120f10fdc9230192d6b157