Use safeArrForEach for when the function might mutate the array
Previously, only every other transport in eg the pendingTransports array
was getting disconnected by disconnectAllTransports, as each
disconnection (in deactivateTransport) was mutating pendingTransports to
remove itself. This would be ok if the removal used `delete`, but it
(meaning Utils.arrDeleteValue) used slice, which renumbers as it goes,
so array iteration gets messed up.