If upgrade sync fails, we’re in an unknown state, so disconnect the transports
If there's been some problem with syncing, we have a problem -- we
can't just fall back on the old transport, as we don't know whether
realtime got the sync -- if it did, the old transport is no longer
valid. To be safe, we disconnect both and start again from scratch.