Redo transport error handling for RTN15i and RTN15h compliance
From now on the protocol message action is determines whether the error
is fatal to the connection. Ably will eventually send full protocol
messages to comet too (rather than `{error: {}}` structs, which is what
it's been sending up to now).

We are now more rigorous about events: a 'failed' event from the
transport implies an issue that is fatal to the connection. For all else
(including all things we called transport#abort() for) we now use
'disconnected'. (We can then also get rid of transport#abort as it
duplicates transport#disconnect).

ConnectionManager is no longer responsible for determining whether a
problem is fatal to the connection. For problems pre-connection, where
the event is not being watched for directly, the error returned in the
tryConnect callback is now a struct with the event name in it