Remove realtimechannel.autonomousAttach()
When autonomousAttach was added in 0.9, its purpose was to emit an error event
with a specific code, that was to be emitted only for attaches triggered
autonomous or implicitly (i.e. not due to the dev calling attach()).
https://github.com/ably/ably-js/commit/081a1784#diff-ee8e6cae22c16e2972f2b341a26d2ed0R193

In a later version of the 0.9 spec, we changed our minds about the error
event and got rid of it.

But autonomousAttach managed to hang on, a little sadly, a bit
vestigial, just pottering around doing a bit of logging now and again,
resigned to be deleted once someone took notice of its continued
existence.