ConnectionManager: ensure we can stop a connection attempt in its early stages
We want disconnectAllTransports to be able to stop any in-progress
connection, even before it gets to the stage of having a pending (or
even a proposed) transport that it can dispose() of. So we check that
it's still current after any early async stage, up until the stage that
is synchronous with instantiating a transport.