Remove buggy callback argument from channel#unsubscribe()
Was throwing an error in the case of a failed channel.

I don't like the idea of a method taking a callback that is only ever
called as a failure condition, and otherwise remains uncalled. And it's
inconsistent with the convention of how callbacks are used in the rest
of the library, which is that if a method takes a callback it is (a)
guaranteed to be called eventually (on either success or failure), and
(b) the method returns a promise if no callback is supplied. Neither of
which is true for unsubscribe().

And there's no reason to conceptualize unsubscribe() as failing if
called on a failed channel anyway, we can still remove the listener
perfectly well. The channel is not going to receive any more messages
anyway, but then that's also true of a detached channel

And nothing in the spec requires it

with '#' will be ignored, and an empty message aborts the commit.