Fix Utils.randomString: make BufferUtils support a much wider range of binary types as input
Utils.randomString on node was calling base64Encode with an ArrayBuffer,
but it only worked with a buffer. Fixed by allowing both browser and
node versions of the various buffertools to take all the common binary
types on each platform.

(And added toBuffer that generates Buffers on node and Uint8Arrays on
browsers, for use with msgpack-encoding -- the node version of
msgpack-js, specifically bops, doesn't like being fed an arraybuffer)