Use "npm run ..." to ensure we use the local version of grunt-cli
grunt-cli 1.3.2 changes how args are passed in a way that breaks
'grunt test:nodeunit --test ...'. This is fine as we specify 1.2.0. But
just calling `grunt` will use whatever grunt is in the user's $PATH,
which we have no control over. So switch to `npm run` commands that
explicitly reference the local repo grunt-cli.

(nb: for grunt commands that aren't explicitly referenced, need to pass
 `--` after `npm run grunt` and before argument passed to grunt, or the
 arguments are passed to `npm` rather than `grunt`)