RealtimeChannel: add last-resort errorinfos if attach or detach fail
for a reason that isn't reflected in the statechange or
connection state.

This should never happen. But if it does due to another lib bug (like
the one we just fixed in the previous commit, RTL5i), at least we
won't incorrectly resolve the attach/detach due to the err being null.