Remove Node’s AgentOptions type from ably.d.ts
This reverts a change introduced in 9b69707. In code review, Quintin
expressed concerns about introducing a dependency on typings for a
third-party library (i.e. the Node standard library) in our typings. I’m
a bit out of my depth here in terms of coming up with a better solution,
so I’m just going to revert that change.