Add placeholder documentation for undocumented reflections
This adds placeholder documentation for all of the reflection types in
TypeDoc’s default value for the requiredToBeDocumented option:

https://github.com/TypeStrong/typedoc/blob/80fa823242746145d2439efd53b1b55430068d3a/src/lib/utils/options/sources/typedoc.ts#L424-L435

> "Enum",
> "EnumMember",
> "Variable",
> "Function",
> "Class",
> "Interface",
> "Property",
> "Method",
> "Accessor",
> "TypeAlias",

This is so I can enable TypeDoc’s notDocumented validation. There
doesn’t seem to be any more elegant way to handle this situation. I had
thought that perhaps there would be some ability to specify a list of
undocumented reflections to ignore, but there isn’t.

As for _why_ I want to throw an error for undocumented reflections, I
think that it’ll keep us on our toes about making sure we document new
things that we add to the SDK. And if we really can’t add documentation
in the moment that we add something new, it’ll force us to add a
placeholder comment and hopefully that will trigger us to create a
GitHub issue to make sure the documentation gets added.

We’ll replace all of these placeholders with proper documentation, which
the tech writers are currently working on as part of an EdX hill.