Extract ClientOptions.recover callback to a separate type
When I tried to add a documentation comment to this inline type,
Prettier would move the comment to a slightly different position,
meaning that the documentation comment no longer applied to this type.

So instead of getting into the weeds of understanding why Prettier
behaves like that, I’m just pulling the inline type out into an explicit
type.