Add a landing page to TypeDoc-generated docs
By default, TypeDoc uses the content of our README.md as the start page
for its generated documentation. But I don’t think it’s particularly
suitable. Most of the links in the README take the user to external
sites (e.g. documentation on ably.com), whereas I think we want to be
primarily explaining to the reader how to use the documentation they’re
currently looking at – i.e. internal links to entry points to the
library, that sort of thing.

This content is just a placeholder, and proper content will come from
tech writers in EDX-155 and we’ll add it here in EDX-207. This content
will probably also include things of a “you’re viewing the TypeDoc
documentation but we also have some lovely custom documentation on
ably.com; here’s a link if you think you’d prefer that” nature.