Change mechanism used for re-exporting Types namespace
This is groundwork for generating separate documentation for the
Promise-based version of the library (that which a user gets when they
import `ably/promises`). I’ll explain why we want to do this in a
subsequent commit message.

In addition to ably.d.ts, we have a separate file promises.d.ts, which
is the TypeScript declarations file used when a user imports
`ably/promises`.

This file is responsible for re-exporting the `Types` namespace from
ably.d.ts, as well as setting the top-level `Rest` and `Realtime`
exports to be the promise-based version of these classes.

Our current approach for this file has explicit re-exports for each of
the types in the Types namespace. As well as the currently explained
problem of this file needing to be manually kept in sync with ably.d.ts,
another problem is that none of these re-exports have any documentation
comments, meaning that:

1. TypeDoc documentation generated from promises.d.ts will not have any
   descriptions.
2. A user who imports `ably/promises` will not get any useful
   IntelliSense documentation in their code editor.

So, I’ve tried a different approach for re-exporting the Types
namespace. To be very honest, I don’t really understand why this
approach works, but it seems to. I pretty much arrived at it by trial
and error. I saw the documentation on namespace aliases
(https://www.typescriptlang.org/docs/handbook/namespaces.html#aliases)
and wondered what would happen if I just tried to export that object.

I also tried export = { ...Types, <the Rest and Realtime types> } but
that introduced a weird namespace named 'export = ' in my TypeDoc docs
🤷.

I did my best to check that this change doesn’t affect the experience of
using the library. That is, I tested it locally by `npm install <path to
my local ably-js repo>` in a new repo, checking that `tsc` works for
compiling a file that imports `ably/promises`, and checking that
IntelliSense works in Visual Studio Code.