Accept audited canonical docstrings that match the legacy ones
(The stuff I marked in 7fc0c55.)

Here, “audited” just means that I read the strings and that they don’t
contradict my knowledge of the ably-JS SDK or my broader understanding
of how our SDKs work. Definitely not a proper proof-read (for style or
content), which is outside the scope of this pull request.