Accept audited canonical docstrings
As in ecfb3fe, “audited” here means that I read the docstrings and
didn’t see any obvious contradictions to my understanding of how ably-js
works or how our SDKs more broadly work. It also means that the
canonical docstrings do not obviously contradict the legacy docstrings.

It does not, however, mean that the canonical docstrings are in any way
“better” than the legacy ones. For example, there are definitely places
where the legacy docstrings are far more detailed than the canonical
ones. But I’ve assumed that to be intentional on the part of the
canonical docstrings, and even if it’s not, addressing that isn’t within
the scope of this work.