Temporary fix (needs upstreaming to IDL and then canonical): Fix default value for ClientOptions#fallbackHosts
The IDL and canonical table say that this property is of non-nullable
type String[], but yet has default value nil, which can’t be right.

The actual default value of this property is defined by (at least)
RSC15h and RSC15i. The default value I’ve added here is just a
placeholder, we need to read the IDL carefully and add the correct value
here.

We need to first fix this default value in the IDL, then update the
canonical docstrings.