ConnectionManager: Fix uprade bug that could lead to an indefinitely sync-pending transport
If we start an upgrade, but then the previous active transport
disconnects by the time the upgrade transport connects, we just want to
activate it immediately. But the previous logic was sending you down the
activateTransport path in that case, and the problem with that is that
the library only sends a sync() message to the server in the
scheduleTransportActivation path. So the server is left waiting for a
sync that will never come, and stays in upgrade-pending limbo.

Fixed by just using the scheduleTransportActivation path, which can cope
fine with the lack of an active protocol (though made a small tweak to
that logic too)