Remove options param from Realtime channel publish methods
As described by feature spec RSL1l:

> The publish(Message) and publish(Message[]) forms of the method should
> take an extra Dict<String, Stringifiable> argument. These parameters
> should be encoded using normal querystring-encoding and sent as part of
> the query string of the REST publish. (Stringifiable is defined in
> RTC1f)

That is, the _REST_ publish methods should accept an additional
dictionary parameter, not the Realtime ones.

And indeed, in src/common/lib/client/realtimechannel.ts#L190-L242,
you’ll see that the Realtime publish method doesn’t do anything with the
options argument.

This is technically a breaking change, but discussed it with Owen and he
thinks it’s okay to do it without a major version release:

> I think we should go ahead and change it based on:
>
> - It's only a breaking change to TS definitions so it's not gonna affect
>   CDN users and break someone;'s app in production - they'll need to
>   pull down the new version and compile again to see an error.
> - If anyone's affected by this (and I'd be surprised if anyone was) then
>   it's probably good to make them aware of it since they're probably
>   mistaken about what it's actually doing.

This partly addresses #1304 (still need to check the REST signatures are
correct).