realtime: add subscription filters support
This commit adds a new getDerived method for obtaining an instance of a "derived
channel" which applies a filter to messages in the upstream channel.

The channel filter is specified via a base64 encoded JMESPath expression
which is given as the value of a new channel qualifier called "filter".

This feature is currently only available in preview and a feature flag
must be enabled on your Ably account to use it.