Update outdated references to ES3
Whilst working on the webpack 5 upgrade, it wasn’t clear to me which
version of ECMAScript we’re actually targeting. But f2014f7 ("Update to
ES5") and #839 ("Now we no longer advertise support for IE8 (or ES3)")
suggested that we’ve dropped ES3 support. Owen subsequently confirmed
this [1]. So let’s update the outdated references to ES3.

[1] https://github.com/ably/ably-js/pull/1201#issuecomment-1520018754