Remove support for JSONP
As #1198 says, "Internal queries show that no one is using JSONP and due
to this, we have decided to remove it from the library in the next major
version update."

Note that in both the web and nodejs platforms,
Defaults.baseTransportOrder and Defaults.transportPreferenceOrder now
have identical contents. I considered combining them into a single
property, but I couldn’t think of a way of expressing the two separate
concepts that I think these properties are trying to express — that is,
"transport most likely to be supported" and "transport we’d most like to
use".  So I’ve kept them separate.

Resolves #1198.