Add spec-conforming logging API
Cherry-picked from commit eaee6f6 in the integration/v2 branch, but
preserving the old API with a deprecation warning. Message of that
commit:

> Conform to spec for logging configuration
>
> Replace the ClientOptions.log property with separate logLevel and
> logHandler properties, to conform with TO3b and TO3c.
>
> Resolves #642.

Decided to add the new API to v1 so that we can add a deprecation
warning to the old one before removing it in v2. This will then allow
the first step of our v2 migration guide to simply be “upgrade to the
latest version of v1 and address all the deprecation warnings”.

Part of #1666.