Remove ProtocolMessage.connectionKey
The library no longer uses this property; rather, it uses the property
of the same name on ConnectionDetails.

This property was removed from the features spec in [1].

This work was scheduled as part of ably-js v2, but I believe we can
remove it in v1 given that ProtocolMessage is not part of the public API
of the library (i.e. it doesn’t appear in ably.d.ts).

Resolves #365.

[1] https://github.com/ably/specification/commit/2d5462cd920614de046656379f79c69667eddbbf