Move remaining DeprecatedClientOptions into new type
These are properties which have been removed from the public API of the
SDK, but which internal and test code still use for modifying the
behaviour of the SDK via its public constructors.

In the case where users were specifying these properties when using v1
of the library, we want these properties to cease to have any effect
upon upgrading to v2. Hence we move them into a new `internal:`
namespace. I hope that it’s safe to assume that users are not going to
pass a set of options that for some reason has a property named
`internal`.

Part of #1197.