Rename IPlatform to IPlatformConfig
Whilst trying to understand how the platform configuration mechanism
works I noticed that we have multiple things called `Platform` and found
it rather confusing.

The new naming matches how these objects are being used in the
platforms’ index.ts files — that is, imported as `Config` and used as
the value of `Platform.Config`.