Remove mention of Platform from README
It’s not part of the public API of the SDK (e.g. doesn’t appear in
ably.d.ts). That does now lead to a question — now that it’s been
accidentally documented, do we need to remember to maintain the
Ably.Realtime.Platform.Crypto property or are we at liberty to rename /
change as we see fit without considering it a breaking change?

In response to the above question, Owen said:

> Good question. I'd say we can still change it - anyone who came across
> this snippet would likely have noticed that something's awry since
> `Realtime.Platform` isn't present in the type declarations.