Convert platform crypto.js files to TypeScript
My eventual aim is to convert Crypto and CipherParams to classes, but
I’m going to do that in a separate commit so that the conversion to
TypeScript is reviewable without being lost in a sea of indentation
changes. The underscored interfaces that this commit introduces are
temporary and will be removed upon conversion to classes.

Resolves #1252.