Make ErrorInfo implement the public interface of same name
This requires us to make its `code` and `statusCode` properties
mandatory and non-nullable.

I initially attempted to implement this by simply adding the appropriate
arguments (values pretty arbitrarily chosen) to all of the ErrorInfo
constructor calls that didn’t currently have them. However, this caused
the authUrl_403_previously_active test to stop receiving its expected
status code, and looking into it, it seems like we are internally using
an ErrorInfo with `code === null` to communicate something (?) which
then gets picked up by auth.ts’s normaliseAuthcallbackError function.

So I decided to change approach, introducing a new PartialErrorInfo
class which behaves like the previous ErrorInfo, to be used when we
don’t have a `code` or `statusCode` to provide. It was the
lightest-touch approach I could think of.

Some current usage of the ErrorInfo type gets replaced with the
IPartialErrorInfo interface, which is satisfied by ErrorInfo and
PartialErrorInfo.

(The setPrototypeOf stuff in PartialErrorInfo’s constructor is just
copied from that of ErrorInfo — I don’t really know what it does.)

Preparation for converting the platform crypto.js files to TypeScript
(#1252).