Make ErrorInfo class’s `code` non-nullable
For consistency with the ErrorInfo type in ably.d.ts (and hence the IDL
in the features spec).

I wasn’t really sure what values to use here, so I went with the
generic-sounding 50000 ("internal error" according to [1]), except for a
timeout where I used 50003 ("timeout error").

Preparation for converting the platform crypto.js files to TypeScript
(#1252).

[1] https://github.com/ably/ably-common/blob/de873312b58144dd3da61c5281aceafb296071cc/protocol/errors.json