Create ICryptoStatic and ICipher interfaces
These describe the interface which is currently exposed by the Crypto
and CBCCipher "classes" exposed by the Node and web crypto.js files. I
wrote them based on the implementation of those classes as well as
looking at how they are used within the codebase and tests.

Preparation for converting platforms’ crypto.js files to TypeScript
(#1252).