Make IBufferUtils generic
This makes it easier to understand, when reading a given platform’s
BufferUtils implementation file, which types they actually need to be
able to deal with.

This doesn’t make any changes at a global level, where the
Platform.BufferUtils object’s method signatures still claim to be able
to handle any buffer type. To resolve this would probably need some
changes to the way we build the library (so that the relevant buffer
types are statically known at build time).