Make IBufferUtils generic
This makes it easier to understand, when reading a given platform’s
BufferUtils implementation file, which types it actually needs to be
able to deal with. It also allows us to remove some type assertions
inside the Node BufferUtils implementation.

This also allows other components to opt in to narrowed type information
when interacting with an IBufferUtils instance, which I intend to do
when converting the platform crypto.js files to TypeScript in #1252.