Tell Prettier to ignore ably-common submodule
This submodule contains some JSON files that don’t conform to Prettier
style. We want `format:check` to not fail because of them, and `format`
to not modify them. (I guess that the `format:check` invocation in the
`check` GitHub workflow has been succeeding because that workflow isn’t
configured to check out submodules.)