Fix Node’s Crypto.getDefaultParams with ArrayBuffer key
getDefaultParams is meant to take an argument of type ably.d.ts’s
Types.CipherParamOptions, whose `key` property is of type `ArrayBuffer |
Uint8Array | string`.

However, the test case that I’ve added here, which attempts to call
`getDefaultParams` with a key of type `ArrayBuffer`, was failing due to
the key length check in `validateCipherParams` failing.

But I believe then even beyond that, attempting to use the resulting
CipherParams to encrypt or decrypt a message would fail on Node < 15,
since support for passing Node’s `crypto.createCipheriv` a key of type
`ArrayBuffer` was only added in Node 15.0.0.

Resolves #1278.