Align signature of Message.fromEncoded* with ably.d.ts
This requires us to remove the checking of the channelCipher property,
which does not exist on the public ChannelOptions type. As far as I can
tell, this check is redundant since normalizeCipherOptions will only
ever be called with a channel options object passed via the public API,
which would not expect to have a channelCipher property.