Added new action to push JS SDK into new CDN bucket
We are moving CDN S3 bucket into a new account, and with a new name (cdn.ably.com). To enable the migration, we need this GHA to temporarily push builds into both the old and new buckets. I have set up a new AWS Role in the SDK account that should be used to publish builds to the new CDN bucket