Always use `instanceof` for checking if something is an ArrayBuffer
Gives richer type information to compiler (i.e. whether the object was
created with the ArrayBuffer constructor instead of whether it just
conforms to the ArrayBuffer interface), which I believe a type guard
can't express.

I’m doing this because the the compiler seemed to consider a TypedArray
to satisfy the ArrayBuffer interface:

> const foo: TypedArray = new Uint8Array();
> const bar: ArrayBuffer = foo; // This compiles, weird!

The effect of this was that, given an object of TypeScript type
`ArrayBuffer | TypedArray`, then, after performing an `is ArrayBuffer`
type guard that returned false, the compiler believed the object to be
of type `never` (when in fact it could still be a TypedArray).