Describe the types of key returned by generateRandomKey
My understanding of the intended use case of this method is that it
allows a user to generate a key which they can then share with other
parties through some side channel in order to establish encrypted
communication with those parties. Hence, the output of generateRandomKey
should not be an opaque value — the user needs to know how to interact
with this value so that they can serialise it.

Resolves #1296.