Update Realtime tests to use SDK’s Promise-based API
I’ve chosen to do this with the lightest touch possible — that is,
maintaining the tests’ callback-based approach and simply bridging the
SDK’s Promise-based API back to callbacks. I did this for the sake of
reviewability and not accidentally changing the behaviour of the tests
in some subtle way that I’d then have to put time into understanding.

It would be good to, at some point, update the structure of the tests to
use `async` / `await`, to improve readability and to make them reflect
how the users actually interact with the Promise-based API in the real
world. I’ve split out into the separate task #1348.

Note that I haven’t made any changes to the calls to EventEmitter’s
`once` or `whenState` methods — we’re going to keep the callback-based
versions of those methods, since there is no Promise equivalent of being
able to turn them off using `off`.

Resolves #1213.