Remove Realtime < Rest inheritance in public API
I want to merge the following classes together:

- RestBase and RestPromise into a class named Rest
- RealtimeBase and RealtimePromise into a class named Realtime

However, I need to decide what to do about the inheritance of
RealtimeBase from RestBase. The obvious thing to do would be to make
Realtime inherit from Rest, but this won’t work because
RealtimeChannel’s type declaration does not include a `status()`
function and hence Realtime.channels cannot satisfy the type of
Rest.channels.

So, since the IDL does not mention any inheritance relation between the
REST and Realtime classes, I’m going to sever this connection in the
type declarations. (We can consider the fact that _internally_ there is
inheritance to be an implementation detail.)