Implement REST token revocation
As described by spec at commit 18ef967. Documentation based on
sdk-api-reference repo at commit 652ef2f.

Note that the tests which revoke tokens do so using a random clientId.
This is to prevent the revocations performed in one test from having an
effect on tokens issued in a subsequent test. (I observed this
happening, and Simon confirmed that it's possible for a token issued
very shortly after a revocation to be affected by that revocation.)

Resolves #989.