Make crypto functionality tree-shakable
We move the crypto functionality into a tree-shakable Crypto module.

Then, we split the decodeMessage* functions introduced in ed90ac8 in
two; we introduce new decodeEncryptedMessage* variants which import the
Crypto module and are hence able to decrypt encrypted messages, and we
change the existing functions to not import the Crypto module and to
fail if they are given cipher options.

Resolves #1396.