Update class hierarchy in preparation for tree-shakability
We rename the Rest class to BaseClient, introduce a new class BaseRest,
and add the prefix "Base" to the name of the Realtime class.

The BaseRest and BaseRealtime classes will be exported by the upcoming
tree-shakable variant of the library. The Default* classes become the
exports of the current (i.e. non tree-shakable) variant.

As we make the library tree-shakable, we will extract functionality from
the BaseClient and BaseRealtime classes and move them into modules
exported by the tree-shakable variant of the library. Whatever
functionality we remove from the Base* classes will be reintroduced in
their Default* counterparts.

The purpose of the BaseRest class is probably not immediately obvious,
since it adds no functionality on top of BaseClient. However, upon
introducing the tree-shakable version of the library, which will split
the REST functionality into a separate module that Realtime can
optionally use, we will use the BaseRest class to ensure that even in
the tree-shakable version of the library, the BaseRest class always
includes REST functionality.

Resolves #1415.

Co-authored-by: Owen Pearson <owen.pearson@ably.com>