Remove Message-related static things in tree-shakable library
This removes the static BaseClient.Message property and the static
Message.{fromEncoded, fromEncodedArray} methods.

The motivation, and the approach taken for the tree-shakable and non
tree-shakable versions of the library, are as in fcda3da.

Note that instead of fromEncoded and fromEncodedArray, the standalone
functions are named decodeMessage and decodeMessages — after some
discussion we decided that this naming was more consistent with the kind
of naming used for standalone functions in JavaScript libraries.