Remove ability to pass array of event names to EventEmitter.prototype.once
This functionality is implemented by wrapping the listener argument in
another listener.  This means that the event emitter does not hold a
reference to the listener argument (other than that held indirectly
through the wrapper) and so it is not possible to remove this listener
using `off(..., listener)`.

The client library specification does not specify a version of `once`
which accepts an array of event names, and we do not advertise it as
part of the public API. So, I think the simplest thing is to remove this
functionality.

Resolves #1452.