fix: improve channel options reattachment logic
This internal change makes the client more selective about which options
provided to `channels.get` will be interpreted as requiring
reattachment. More precisely, the channel params and modes will now be
shallow compared to existing params and modes, and only if they differ
will the call the `chnanels.get` throw an error (, provided of course
that the channel is already attached or attaching).