Add typings for tree-shakable version of library
I wasn’t sure of the best way to approach the typings of the modules
(`Rest` etc). They should be opaque to the user (they shouldn’t be
interacting with them directly and we want be free to change their
interface in the future).

There is an open issue to add support for opaque types to TypeScript
[1], and people have suggested various sorts of ways of approximating
them, which revolve around the use of `unique symbol` declarations.
However, I don’t fully understand these solutions and so thought it best
not to include them in our public API. So, for now, let’s just use
`unknown`, the same way as we do for `CipherParams.key`.

Resolves #1442.

[1] https://github.com/Microsoft/Typescript/issues/202