Upgrade Playwright to 1.39.0
This requires us to now explicitly install the browsers (see [1]).

I’m doing this upgrade not because it’s necessary but because NPM tried
to make it for me when resolving a package-lock.json merge conflict in
an upcoming commit. Given that the upgrade involves this CI change,
thought it best to do it as a separate commit so it doesn’t get lost
inside a merge commit.

[1] https://playwright.dev/docs/release-notes#breaking-changes-playwright-no-longer-downloads-browsers-automatically