Add documentation for the modular version of the library
I imagine there’s plenty of room for improvement here, but it’s a start.

Hopefully a tech writer will be able to have a look at it before we
release, too.

For handling TypeDoc documentation for the two different variants of the
library, I’ve resurrected the approach that we used when we had
callback-based and Promise-based variants of the library (removed in
2a2ed49).

Resolves #1443.